Refactor: Move timestamp to shared
